Create a separate finagle-h2 module containing only the client and server
Currently, the entire HTTP/2 implementation is packaged in the router-h2
package. It's more consistent with other protocols to split out the core codec
functionality, so this changes creates a finagle-h2 package that isolates
com.twitter.finagle.buoyant.

router-h2 now only exposes io.buoyant.router.{H2,h2}.
