+++ finagle/README.md
@@ -1,0 +1,4 @@
+# linkerd: finagle extensions #
+
+This directory contains protocol extensions to Finagle published under
+the `io.buoyant` organization.
