+++ finagle/h2/README.md
@@ -1,0 +1,4 @@
+# io.buoyant:finagle-h2 #
+
+Status: **Pre-production** -- currently testing for production-readiness
+
