+++ finagle/h2/src/main/scala/com/twitter/finagle/buoyant/h2/Stream.scala
