+++ finagle/h2/src/main/scala/com/twitter/finagle/buoyant/h2/netty4/Netty4H2Transporter.scala
@@ -6,7 +6,6 @@ import com.twitter.finagle.client.Transporter
 import com.twitter.finagle.netty4.Netty4Transporter
 import com.twitter.finagle.netty4.buoyant.BufferingConnectDelay
 import com.twitter.finagle.transport.{TlsConfig, Transport}
-import io.buoyant.router.H2
 import io.netty.channel.ChannelPipeline
 import io.netty.handler.codec.http2.{Http2FrameCodec, Http2Frame}
