+++ router/h2/src/main/scala/io/buoyant/router/H2.scala
@@ -1,31 +1,39 @@
 package io.buoyant.router
 
 import com.twitter.finagle._
-import com.twitter.finagle.buoyant.Dst
+import com.twitter.finagle.buoyant.{Dst, H2 => FinagleH2}
 import com.twitter.finagle.buoyant.h2.{Request, Response, Reset}
-import com.twitter.finagle.buoyant.h2.netty4._
-import com.twitter.finagle.buoyant.h2.param._
 import com.twitter.finagle.param
 import com.twitter.finagle.client.{StackClient, StdStackClient, Transporter}
 import com.twitter.finagle.server.{Listener, StackServer, StdStackServer}
 import com.twitter.finagle.service.StatsFilter
 import com.twitter.finagle.transport.Transport
 import com.twitter.util.{Closable, Future}
-import io.netty.handler.codec.http2.Http2Frame
 import java.net.SocketAddress
 
-object H2 extends Client[Request, Response]
-  with Router[Request, Response]
-  with Server[Request, Response] {
+object H2 extends Client[Request, Response] with Server[Request, Response]
+  with Router[Request, Response] {
 
-  private[this] val log = com.twitter.logging.Logger.get(getClass.getName)
+  val client = FinagleH2.client
 
-  private[this]type Http2FrameTransporter = Transporter[Http2Frame, Http2Frame]
-  private[this]type Http2FrameTransport = Transport[Http2Frame, Http2Frame]
+  def newService(dest: Name, label: String): Service[Request, Response] =
+    client.newService(dest, label)
 
-  /**
-   * Clients and servers may accumulate data frames into a single
-   * combined data frame.
+  def newClient(dest: Name, label: String): ServiceFactory[Request, Response] =
+    client.newClient(dest, label)
+
+  object Server {
+    val newStack: Stack[ServiceFactory[Request, Response]] = FinagleH2.Server.newStack
+      .insertAfter(StackServer.Role.protoTracing, h2.ProxyRewriteFilter.module)
+  }
+
+  val server = FinagleH2.server.withStack(Server.newStack)
+
+  def serve(addr: SocketAddress, service: ServiceFactory[Request, Response]): ListeningServer =
+    server.serve(addr, service)
+
+  /*
+   * Router
    */
 
   case class Identifier(mk: Stack.Params => RoutingFactory.Identifier[Request])
@@ -36,59 +44,9 @@ object H2 extends Client[Request, Response]
     val default = Identifier(params => nil)
   }
 
-  /*
-   * Client
-   */
-
-  object Client {
-    val newStack: Stack[ServiceFactory[Request, Response]] =
-      StackClient.newStack
-        .insertAfter(StatsFilter.role, h2.StreamStatsFilter.module)
-
-    val defaultParams = StackClient.defaultParams +
-      param.ProtocolLibrary("h2")
-  }
-
-  case class Client(
-    stack: Stack[ServiceFactory[Request, Response]] = Client.newStack,
-    params: Stack.Params = Client.defaultParams
-  ) extends StdStackClient[Request, Response, Client] {
-
-    protected type In = Http2Frame
-    protected type Out = Http2Frame
-
-    protected def newTransporter(): Http2FrameTransporter =
-      Netty4H2Transporter.mk(params)
-
-    protected def copy1(
-      stack: Stack[ServiceFactory[Request, Response]] = this.stack,
-      params: Stack.Params = this.params
-    ): Client = copy(stack, params)
-
-    private[this] lazy val param.Stats(statsReceiver) = params[param.Stats]
-    private[this] lazy val streamStats =
-      new Netty4StreamTransport.StatsReceiver(statsReceiver.scope("stream"))
-
-    protected def newDispatcher(trans: Http2FrameTransport): Service[Request, Response] =
-      new Netty4ClientDispatcher(trans, streamStats)
-  }
-
-  val client = Client()
-
-  def newService(dest: Name, label: String): Service[Request, Response] =
-    client.newService(dest, label)
-
-  def newClient(dest: Name, label: String): ServiceFactory[Request, Response] =
-    client.newClient(dest, label)
-
-  /*
-   * Router
-   */
-
   object Router {
     val pathStack: Stack[ServiceFactory[Request, Response]] = {
-      val stk = StackRouter.newPathStack
-        .insertAfter(StatsFilter.role, h2.StreamStatsFilter.module)
+      val stk = StackRouter.newPathStack[Request, Response]
       h2.ViaHeaderFilter.module +: stk
     }
 
@@ -123,55 +81,4 @@ object H2 extends Client[Request, Response]
 
   def factory(): ServiceFactory[Request, Response] =
     router.factory()
-
-  /*
-   * Server
-   */
-
-  object Server {
-    val newStack: Stack[ServiceFactory[Request, Response]] = StackServer.newStack
-      .insertAfter(StackServer.Role.protoTracing, h2.ProxyRewriteFilter.module)
-      .insertAfter(StatsFilter.role, h2.StreamStatsFilter.module)
-
-    val defaultParams = StackServer.defaultParams +
-      param.ProtocolLibrary("h2")
-  }
-
-  case class Server(
-    stack: Stack[ServiceFactory[Request, Response]] = Server.newStack,
-    params: Stack.Params = Server.defaultParams
-  ) extends StdStackServer[Request, Response, Server] {
-
-    protected type In = Http2Frame
-    protected type Out = Http2Frame
-
-    protected def copy1(
-      stack: Stack[ServiceFactory[Request, Response]] = this.stack,
-      params: Stack.Params = this.params
-    ): Server = copy(stack, params)
-
-    /**
-     * Creates a Listener that creates a new Transport for each
-     * incoming HTTP/2 *stream*.
-     */
-    protected def newListener(): Listener[Http2Frame, Http2Frame] =
-      Netty4H2Listener.mk(params)
-
-    private[this] lazy val statsReceiver = params[param.Stats].statsReceiver
-    private[this] lazy val streamStats =
-      new Netty4StreamTransport.StatsReceiver(statsReceiver.scope("stream"))
-
-    /** A dispatcher is created for each inbound HTTP/2 connection. */
-    protected def newDispatcher(
-      trans: Http2FrameTransport,
-      service: Service[Request, Response]
-    ): Closable = {
-      new Netty4ServerDispatcher(trans, service, streamStats)
-    }
-  }
-
-  val server = Server()
-
-  def serve(addr: SocketAddress, service: ServiceFactory[Request, Response]): ListeningServer =
-    server.serve(addr, service)
 }
